requirements.txt
flask==3.1.0
pdfplumber==0.11.4
openpyxl==3.1.5
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.166.0
